rename gnucsh.py to cli.py to allow installing it
