+++ gnucsh/cli.py
